--- a/src/accept.jl
+++ b/src/accept.jl
@@ -1,0 +1,3 @@
+function always_accept(partition::Partition, graph::BaseGraph)
+    return 1
+end