--- a/src/flip.jl
+++ b/src/flip.jl
@@ -75,9 +75,16 @@ end
 
 function update_partition!(partition::Partition,
                            graph::BaseGraph,
-                           proposal::FlipProposal)
+                           proposal::FlipProposal,
+                           copy_parent::Bool=false)
     """ Updates the Partition with the FlipProposal
     """
+    if copy_parent
+        partition.parent = nothing
+        old_partition = deepcopy(partition)
+        partition.parent = old_partition
+    end
+
     # update district population counts
     partition.dist_populations[proposal.D₁] = proposal.D₁_pop
     partition.dist_populations[proposal.D₂] = proposal.D₂_pop
@@ -99,7 +106,7 @@ function flip_chain(graph::BaseGraph,
                     num_steps::Int,
                     score_keys::Array{String, 1},
                     scores_save_dir::AbstractString="./scores.json",
-                    num_tries::Int=3)
+                    acceptance_fn::Function=always_accept)
     """ Runs a Markov Chain for `num_steps` steps using Flip proposals.
 
         Arguments:
@@ -110,6 +117,9 @@ function flip_chain(graph::BaseGraph,
             score_keys:         list of scores to evaluate plans with
             score_save_dir:     directory of where to store the scores
             num_steps:          Number of steps to run the chain for
+            acceptance_fn:      A function generating a probability in [0, 1]
+                                representing the likelihood of accepting the
+                                proposal
     """
     steps_taken = 0
     all_scores = Array{Dict{String, Any}, 1}()