--- a/test/accept.jl
+++ b/test/accept.jl
@@ -1,4 +1,4 @@
-@testset "pass_acceptance_fn" begin
+@testset "satisfies_acceptance_fn" begin
     graph = BaseGraph(square_grid_filepath, "population", "assignment")
     partition = Partition(square_grid_filepath, graph, "population", "assignment")
 
@@ -9,14 +9,14 @@
     always_fail(p) = 0.0
 
     # partition doesn't have a "parent" defined yet, so the call to
-    # pass_acceptance_fn should fail here. As a reminder, all partitions
+    # satisfies_acceptance_fn should fail here. As a reminder, all partitions
     # passed to an acceptance function should have the parent field assigned.
-    @test_throws AssertionError pass_acceptance_fn(partition, always_accept)
+    @test_throws AssertionError satisfies_acceptance_fn(partition, always_accept)
 
     partition.parent = deepcopy(partition)
-    @test_throws ArgumentError pass_acceptance_fn(partition, inval_acc_fn_1)
-    @test_throws ArgumentError pass_acceptance_fn(partition, inval_acc_fn_2)
-    @test_throws MethodError pass_acceptance_fn(partition, inval_acc_fn_3)
-    @test pass_acceptance_fn(partition, always_pass) == true
-    @test pass_acceptance_fn(partition, always_fail) == false
+    @test_throws ArgumentError satisfies_acceptance_fn(partition, inval_acc_fn_1)
+    @test_throws ArgumentError satisfies_acceptance_fn(partition, inval_acc_fn_2)
+    @test_throws MethodError satisfies_acceptance_fn(partition, inval_acc_fn_3)
+    @test satisfies_acceptance_fn(partition, always_pass) == true
+    @test satisfies_acceptance_fn(partition, always_fail) == false
 end