In this PR, I implement functionality that allows the user to pass in a custom acceptance function to `recom_chain` and `flip_chain.` Acceptance functions allow users to compare an updated partition (i.e., the next step in the chain) to the original partition and decide a probability with which to accept or reject the updated partition. (The probability may be 0 or 1). If the new partition is rejected, the step counter is incremented and the new partition is discarded (i.e., a self-loop in the Markov chain).

- `src/accept.jl` contains a standard acceptance function, `always_accept`, as well as `pass_acceptance_fn`, which takes in the user function for generating the acceptance probability, runs a single Bernoulli trial with that probability, and returns `true` or `false`. I implement checks to ensure various conditions about the user-defined acceptance function hold true, such as whether it returns a probability in the range [0,1].
- `src/flip.jl` and `src/recom.jl` were modified to take as an argument an optional `acceptance_fn` parameter
- Test: `tests/accept.jl` tests `pass_acceptance_fn`

**Performance Impact**
_Without acceptance fxns:_
Recom chain, PA VTD, 100 steps:  6.494099 seconds (2.40 M allocations: 185.112 MiB, 2.90% gc time)
Flip chain, PA VTD, 10000 steps: 32.864040 seconds (180.71 M allocations: 3.036 GiB, 1.83% gc time)

_With acceptance fxns:_
Recom chain, PA VTD, 100 steps:  7.317668 seconds (2.41 M allocations: 185.213 MiB, 3.56% gc time)
Recom chain, PA VTD, 100 steps, with toy acceptance fxn (always returns 0.5):  5.672623 seconds (2.53 M allocations: 223.098 MiB, 3.04% gc time)
Flip chain, PA VTD, 10000 steps:  32.828002 seconds (185.48 M allocations: 3.106 GiB, 1.76% gc time)
Flip chain, PA VTD, 10000 steps, with toy acceptance fxn (always returns 0.5):  38.601872 seconds (183.70 M allocations: 5.836 GiB, 3.21% gc time)

**Conclusion**: No significant time/memory impact on on code without acceptance functions, though we do see **substantial memory allocation increases**, likely due to the copying of the partition every time.