@bsuwal 2020-06-26
I love this PR! 

Quick note: It is fine for now but in the future could you do 1000 Recom steps instead of 100? I think the times become more stable then, instead of how right now we are looking at the granularity of 1s, which I find sometimes just varies from run to run..

Otherwise I have no comment, looks good to me.